RunTcrdist3

This function runs the tcrdist3 pipeline on a set of CDR3 sequences in a Seurat Object.

RunTcrdist3(
  seuratObj = NULL,
  metadata = NULL,
  organism = "human",
  formatMetadata = T,
  postFormattingMetadataCsvPath = "./tcrDist3Input.csv",
  chains = c("TRA", "TRB"),
  cleanMetadata = T,
  spikeInDataframe = NULL,
  summarizeClones = T,
  imputeCloneNames = T,
  minimumClonesPerSubject = 2,
  rdsOutputPath = "./tcrdist3DistanceMatrices/",
  pythonExecutable = NULL,
  debugTcrdist3 = "True"
)

Arguments

seuratObj

Seurat Object containing TCR information. If NULL, metadata must be provided.

metadata

Data frame containing metadata. If NULL, seuratObj must be provided.

organism

Organism to use for tcrdist3. Default is 'human'.

formatMetadata

Boolean controlling whether to format the metadata for tcrdist3 using the internal FormatMetadataForTcrDist3 function. Default is TRUE.

postFormattingMetadataCsvPath

Path to the output CSV file from FormatMetadataForTcrDist3. Default is './tcrDist3Input.csv'.

chains

Vector of TCR chains to include in the analysis. Default is c("TRA", "TRB").

cleanMetadata

Pass-through boolean controlling whether to clean the metadata by removing rows with NA values or commas in the specified chains. Default is TRUE.

spikeInDataframe

Data frame containing known CDR3s and gene segments to be included in the clustering. Default is NULL.

summarizeClones

Pass-through boolean controlling whether to summarize clones into a frequency (by SubjectId, TRA, TRB, TRA_V, and TRB_J). Default is TRUE.

imputeCloneNames

Pass-through boolean controlling whether to impute clone names if they are missing. Existing clone names will be inherited. Default is TRUE.

minimumClonesPerSubject

Minimum number of clones per subject to include in the analysis. Default is 2.

rdsOutputPath

Path to the output directory for the RDS files containing the distance matrices. Default is "./tcrdist3DistanceMatrices/".

pythonExecutable

Path to the python executable. Default is reticulate::py_exe().

debugTcrdist3

String (to be passed to python and converted to boolean) controlling whether to run tcrdist3 in debug mode. Default is "True".

Examples

if (FALSE) { # \dontrun{
  RunTcrdist3(seuratObj = seuratObj,
              metadata = NULL,
              formatMetadata = T,
              postFormattingMetadataCsvPath = './tcrDist3Input.csv',
              chainsString = "alpha,beta",
              cleanMetadata = T,
              summarizeClones = T,
              imputeCloneNames = T,
              minimumClonesPerSubject = 2,
              rdsOutputPath = "./tcrdist3DistanceMatrices/",
              pythonExecutable = reticulate::py_exe(),
              debugTcrdist3 = "True")

    spikeInDataframe <- data.frame(CloneNames = rep(1:3),
                                 TRA_V = c("TRAV1-2", "TRAV1-2", "TRAV1-2"),
                                 TRA_J = c("TRAJ33", "TRAJ20", "TRAJ33"),
                                 TRA = c("CAVRDSNYQLIW", "CAVSLQDYKLSF", "CAVRDSNYQLIW"),
                                 TRB_V = c("TRBV6-4", "TRBV6-4", "TRBV6-4"),
                                 TRB_J = c("TRBJ1-1", "TRBJ2-1", "TRBJ2-3"),
                                 TRB = c("CASSAAAAAAAAFF", "CASSVVVVVVVVQF", "CASSWWWWWWWWQY"))
} # }